save balance scores as double

+++ src/main/java/DbController.java
@@ -107,9 +107,9 @@ public class DbController {
 
         PreparedStatement pstmt = this.conn.prepareStatement(sql);
         pstmt.setDouble(1, schedule.fairnessScore());
-        pstmt.setInt(2, (int) schedule.balanceScore());
+        pstmt.setDouble(2, schedule.balanceScore());
         pstmt.setDouble(3, schedule.jaevFairnessScore());
-        pstmt.setInt(4, schedule.jaevBalanceScore());
+        pstmt.setDouble(4, schedule.jaevBalanceScore());
         pstmt.execute();
     }
